requirements.txt
python-dotenv>=1.0.0
PyMuPDF>=1.24.0
requests>=2.31.0
transformers>=4.40.0
torch>=2.2.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
langchain>=0.1.14
Pillow>=10.0.0
argon2-cffi>=23.1.0
cryptography>=41.0.3
python-decouple>=3.8
streamlit>=1.32.0
